+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This is a major refactor!
 * awesn1 instead of indispensable-asn1
 * publish javadoc redirect to save space
 * make all provider functions suspend
-* Replace `JwsHeader.Part` and move unprotected-member placement metadata from `JwsHeader` to the JWS representations
+* Remove `JwsHeader.Part`. Unprotected-member placement metadata for `JwsHeader` is now handled by the JWS representations
 * Dependency Updates:
     * Bouncy Castle 1.85
     * multibase 1.3.0